requirements.txt
streamlit==1.27.2
tensorflow==2.12.0
fpdf
pillow
matplotlib
numpy
seaborn
